requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Injector==0.15.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
injector==0.21.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
openai==0.28.0
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.66.1
trycourier==4.4.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
